Finished initial implementation of SQL-to-grt functionality in new parser module.
Parser:
- Added convenience function to test for the current token.
- Added support for parsing: create indices, create tables, create logfile group, create server and create tablespace SQL.

Grammar:
- Reworked index + key naming to follow the server grammar (fqn allowed).
- Optimized key def parsing.
- Fixed a number of smaller issues where we diverted from the parser behavior.

grt structs:
- Added support for missing server features on various db objects (e.g. nodegroup id for a tablesapce).
- Added event db object type.
- Removed separate function and procedure grt classes. We only use routine.
- Regenerated structs classes to accomodate updates for db objects.
- Win: fixed tools solution to properly build wrapper tools. The tools are fully self-contained now and don't need existing WB libs anymore.

Parser services/module:
- Changed all const & for grt Refs so they are used without that. We want them to be copied, not shared.
- Finished initial SQL-to-catalog parser function.

Additional stuff:
- SQL IDE: added reset_references call when we are done with applying live object changes.
- SQL IDE: update TODO item about reset_references.
- SQL IDE: wrapped an OSX-only fix with #ifdef.
- SQL IDE: updated DDL retrieval function to avoid usage of the old SqlFacade and avoid adding unnecessary DELIMITER calls.
- Fixed a number of warnings.
- Object editors: disabled reset_references() call on destruction as this is the wrong place to do that.
- Win: fixed a warning about unintentionally overriding a method by using a different name.
- Win: some small project setting changes (e.g. paths with additional backslash).
- Canvas: fix for an occasional crash when shutting down wrt delayed refreshs (plus a small optimization).

+++ backend/wbpublic/grt/common.cpp
@@ -1,5 +1,5 @@
 /* 
- * Copyright (c) 2007, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2007, 2015, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -402,8 +402,8 @@ namespace bec {
     {
       switch (to)
       {
-        case MoveUp: to = ui_item_index - 1; break; 
-        case MoveDown: to = ui_item_index + 2; break;
+        case MoveUp: to = (int)ui_item_index - 1; break; 
+        case MoveDown: to = (int)ui_item_index + 2; break;
         case MoveTop: 
           to = (grouped && fabric_names[group_name]) ? 1 : 0;
           break;
@@ -418,7 +418,7 @@ namespace bec {
 
       // In case o should point to the last element on the group
       if (to == MoveBottom)
-        to = target_group->count() - 1;
+        to = (int)target_group->count() - 1;
       else 
       {
         // This adjustment is needed because of the way reorder works
@@ -435,7 +435,7 @@ namespace bec {
     {
       // In case o should point to the last element on the entire list
       if (to == MoveBottom)
-        to = names_list.size() - 1;
+        to = (int)names_list.size() - 1;
       // This adjustment is needed because of the way reorder works
       else if ((int)ui_item_index < to)
         to--;
